[fix] don't replace git_url when it is valid.
valid = the static files downloaded by the browser belong to the git_url.

---

If a fork contains exactly the same static files(*),
searxstats don't know which fork to pick.

(*) more precisly: the browser fetchs some static files.
searxstats get the commits related to these files, then the forks.

This commit give priority to the declared git_url,
so the git_url is not replaced.